adding params for eq
the slider component was built together oinkydoinky-like, making it really difficult to get the first and last values required to setup parameters. furthermore, the eq has a positive min value and a negative max value making calculations even more hard to figure out.

this checkin adds a new function called `slider:get_minmax_values` that returns the first and last values of the slider as a table. also, eq params are added that update the eq ui sliders when they are changed. conversely, updating the eq ui sliders updates the params.

fingers crossed there aren't major issues with the code...

the table looks like `{min,max}` where `min` is the first value of the slider and `max` is the last value of the slider.